adding back seed and clear database scripts
